+++ internal/settings/pattern.go
@@ -1,0 +1,8 @@
+package settings
+
+import "regexp"
+
+type PatternConfig struct {
+	RegExp   *regexp.Regexp
+	Template string
+}