+++ internal/webhooks/sonarqube/webhook_test.go
@@ -1,18 +1,29 @@
 package sonarqube
 
 import (
+	"regexp"
 	"testing"
+
+	"gitea-sonarqube-pr-bot/internal/settings"
 
 	"github.com/stretchr/testify/assert"
 )
 
 func TestNewWebhook(t *testing.T) {
+	settings.Pattern = &settings.PatternConfig{
+		RegExp: regexp.MustCompile(`^PR-(\d+)$`),
+	}
+
 	raw := []byte(`{ "serverUrl": "https://example.com/sonarqube", "taskId": "AXouyxDpizdp4B1K", "status": "SUCCESS", "analysedAt": "2021-05-21T12:12:07+0000", "revision": "f84442009c09b1adc278b6aa80a3853419f54007", "changedAt": "2021-05-21T12:12:07+0000", "project": { "key": "pr-bot", "name": "PR Bot", "url": "https://example.com/sonarqube/dashboard?id=pr-bot" }, "branch": { "name": "PR-1337", "type": "PULL_REQUEST", "isMain": false, "url": "https://example.com/sonarqube/dashboard?id=pr-bot&pullRequest=PR-1337" }, "qualityGate": { "name": "PR Bot", "status": "OK", "conditions": [ { "metric": "new_reliability_rating", "operator": "GREATER_THAN", "value": "1", "status": "OK", "errorThreshold": "1" }, { "metric": "new_security_rating", "operator": "GREATER_THAN", "value": "1", "status": "OK", "errorThreshold": "1" }, { "metric": "new_maintainability_rating", "operator": "GREATER_THAN", "value": "1", "status": "OK", "errorThreshold": "1" }, { "metric": "new_security_hotspots_reviewed", "operator": "LESS_THAN", "status": "NO_VALUE", "errorThreshold": "100" } ] }, "properties": {} }`)
 	response, ok := New(raw)
 
 	assert.NotNil(t, response)
 	assert.Equal(t, 1337, response.PRIndex)
 	assert.True(t, ok)
+
+	t.Cleanup(func() {
+		settings.Pattern = nil
+	})
 }
 
 func TestNewWebhookInvalidJSON(t *testing.T) {
@@ -23,8 +34,16 @@ func TestNewWebhookInvalidJSON(t *testing.T) {
 }
 
 func TestNewWebhookInvalidBranchName(t *testing.T) {
+	settings.Pattern = &settings.PatternConfig{
+		RegExp: regexp.MustCompile(`^PR-(\d+)$`),
+	}
+
 	raw := []byte(`{ "serverUrl": "https://example.com/sonarqube", "taskId": "AXouyxDpizdp4B1K", "status": "SUCCESS", "analysedAt": "2021-05-21T12:12:07+0000", "revision": "f84442009c09b1adc278b6aa80a3853419f54007", "changedAt": "2021-05-21T12:12:07+0000", "project": { "key": "pr-bot", "name": "PR Bot", "url": "https://example.com/sonarqube/dashboard?id=pr-bot" }, "branch": { "name": "invalid", "type": "PULL_REQUEST", "isMain": false, "url": "https://example.com/sonarqube/dashboard?id=pr-bot&pullRequest=PR-1337" }, "qualityGate": { "name": "PR Bot", "status": "OK", "conditions": [ { "metric": "new_reliability_rating", "operator": "GREATER_THAN", "value": "1", "status": "OK", "errorThreshold": "1" }, { "metric": "new_security_rating", "operator": "GREATER_THAN", "value": "1", "status": "OK", "errorThreshold": "1" }, { "metric": "new_maintainability_rating", "operator": "GREATER_THAN", "value": "1", "status": "OK", "errorThreshold": "1" }, { "metric": "new_security_hotspots_reviewed", "operator": "LESS_THAN", "status": "NO_VALUE", "errorThreshold": "100" } ] }, "properties": {} }`)
 	_, ok := New(raw)
 
 	assert.False(t, ok)
+
+	t.Cleanup(func() {
+		settings.Pattern = nil
+	})
 }